# Improve `Text.check` performance