--- a/lib/rexml/attribute.rb
+++ b/lib/rexml/attribute.rb
@@ -173,7 +173,7 @@ module REXML
       @element = element
 
       if @normalized
-        Text.check( @normalized, NEEDS_A_SECOND_CHECK, doctype )
+        Text.check( @normalized, NEEDS_A_SECOND_CHECK )
       end
 
       self